+++ project/src/pages/main-screen/main-screen.tsx
@@ -1,8 +1,5 @@
-import React, { useEffect, useState } from 'react';
-import { Hotel } from '../../types/hotel';
+import React, { useEffect } from 'react';
 import { Header } from '../../components/header/header';
-import { CitiesPlaces } from '../../components/cities-places/cities-places';
-import { Map } from '../../components/map/map';
 import classNames from 'classnames';
 import { useAppDispatch, useAppSelector } from '../../hooks';
 import { CitiesList } from '../../components/cities-list/cities-list';
@@ -12,6 +9,7 @@ import { LoaderThreeDots } from '../../components/loader/loader';
 import { getCity } from '../../services/store/slices/city-data/city-data-selectors';
 import { getHotels } from '../../services/store/slices/hotels-data/hotels-data-selectors';
 import { getIsDataLoading } from '../../services/store/slices/user-process/user-process-selectors';
+import { HotelsContainer } from '../../components/hotels-container/hotels-container';
 
 export const MainScreen: React.FunctionComponent = () => {
 
@@ -26,26 +24,12 @@ export const MainScreen: React.FunctionComponent = () => {
     dispatch(fetchHotelsAction());
   }, [dispatch]);
 
-
-  const [selectedHotel, setSelectedHotel] = useState<Hotel | null>(null);
-
   const hotelsByCity = getHotelsByCity(hotels, city);
-
-  const onListItemHover = (listItemId: number | undefined) => {
-    const currentHotel = hotelsByCity.find((item) => item.id === listItemId);
-    setSelectedHotel(currentHotel ? currentHotel : null);
-  };
 
   const mainClass = classNames({
     'page__main': true,
     'page__main--index': true,
     'page__main--index-empty': !hotelsByCity.length
-  });
-
-  const placesContainerClass = classNames({
-    'container': true,
-    'cities__places-container': true,
-    'cities__places-container--empty': !hotelsByCity.length
   });
 
   return (
@@ -58,26 +42,7 @@ export const MainScreen: React.FunctionComponent = () => {
           <div className="tabs">
             <CitiesList />
           </div>
-          <div className="cities">
-            <div className={placesContainerClass}>
-              <CitiesPlaces
-                onListItemHover={onListItemHover}
-              />
-              <div className="cities__right-section">
-                {hotelsByCity.length > 0 &&
-                  <section className="cities__map map">
-                    <Map
-                      selectedHotel={selectedHotel}
-                      hotels={hotelsByCity}
-                      style={{
-                        height: '100%',
-                        width: '512px'
-                      }}
-                    />
-                  </section>}
-              </div>
-            </div>
-          </div>
+          <HotelsContainer></HotelsContainer>
         </main>
       </div>
     </>
